Enforce 7,000-line file limit
